qcam: Add a GUI way to use capture script
Implement an Open Capture Script button which would allow the user to open a Capture Script (*.yaml). When clicked present them with a QFileDialog to allow them to select a single file.

Introduce a queueCount_ to keep track of the requests queued.

Initialize a Script Parser instance when the user selects a valid capture script. At queueRequest() time if script parser has been initialized, then populate the Request::controls() with it at queueRequest time providing the queueCount_.
The queueCount_ is incremented after the getting controls from the parser, so the first request is for frame 0.