qcam: Add a GUI way to use capture script
Implement an Open Capture Script button which would allow the user
to open a Capture Script (*.yaml).
This button has two states Open and Stop.

Open state allows user to load a capture script.
When clicked in open state present them with a QFileDialog
to allow user to select a single file.

Stop state stops the execution of the current capture script.

Introduce a queueCount_ to keep track of the requests queued.

When stopping the execution no count is reset and the
capture continues as it is, this would allow better testing
as we can stop in between script and observe.

Initialize a Script Parser instance when the user selects
a valid capture script.

At queueRequest() time if script parser has been initialized,
then populate the Request::controls() with it at
queueRequest time providing the queueCount_.

The queueCount_ is incremented after the getting controls from
the parser, so the first request is for frame 0.