qcam: Add a GUI way to use capture script
Implement an Open Capture Script button which would allow the user
to open a Capture Script (*.yaml).
This button has two states Open and Stop.

Open state allows user to load a capture script.
When clicked in open state present them with a QFileDialog
to allow user to select a single file.

Stop state stops the execution of the current capture script.

Introduce a queueCount_ to keep track of the requests queued.

When stopping the execution no count is reset and the
capture continues as it is.

Requests are queued with any controls the script matching
the current queueCount_

Signed-off-by: Utkarsh Tiwari <utkarsh02t@gmail.com>