qcam: Add settings Dialog with Control tab
Implement a Settings QDialog and add its buttonon the toolbar.
We only create one instance of the SettingsWindow and its created the
first time user clicks on the button. We check its existence with
QPointer.

This Setting Dialog implements a QTabWidget which resides in a
QVBoxLayout so that it spans the whole QDialog.

The QTabWidget currently only holds the control tab.
This Tab lays out the respective ControlFrames in a QGridLayout listing
two controls in each row.

ControlFrames handles the UI for each respective controls.
It also implements the QVBoxLayout as for future when we want to add
more information in each frame.