qcam: Add GUI way to change control values
Introduce GUI ability to change control values currently only for
boolean control types.

Each of the ControlFrame first identifies the type of control it is and
then accordingly constructs the widget that would help the user to
interact with it.

The current model of connecting the controls change info to the
MaindWindow is as follows.

            +-----------+
            |Main Window|
            +-----^-----+
                  |
                  |
          +-------+-------+
          |Settings Dialog|
          +--^----^------^+
             |    |      |
      +------+    |      +-----+
      |           |            |
+-----+---+  +----+----+  +----+----+
|Various  |  | Tabs    |  |   ...   |
++---+----+  +---+---+-+  ++--+--+--+
 |   |           |   |     |  |  |
 |   |           |   |     |  |  |
 |   |           |   |     |  |  |

Implement a controlLatch mechanism to update thec ontrolList. As the
Settings QDialog and ControlFrame processing the change in control lie
in the same thread there would be no race to access the ControlList.

This allows us to use to shared_ptr to the ControlList
to help us the same function from MaindWindow::queueRequest()
and the controlListChanged() signal.

We clear the controlList_ in MainWindow::queueRequest() and not in the
MainWindow::controlListLatch() because when in controlListLatch() we
need to clear it when we don't have script nor the alerted by the signal.
But what Can happen is that we have controlListLatch() is triggered by the
signal and then immediately called by the queueRequest() which would
clear the controlList and it would not end up getting set.