qcam: ControlFrame: Add Current Value label
Add a QLabel which displays the Current value of the control.
Currently this only supports ControlTypeBool.

At start do not display any Current value as we cannot be sure of the
starting conditions and the controls that have been applied.

When a request completes succesfully completes, we use its ControlList
to update the currentValue.

The current model for sending the controlValue is :

MainWindow -> Settings -> Control Tab -> Control Frame