qcam: Add a GUI way to use capture script
Implement an Open Capture Script button which would allow the user
to open a Capture Script (*.yaml).
This button has two states :
    - Open Capture Script
    - Stop the execution of current capture script

When being clicked in open state, present them with a QFileDialog to
allow user to select a single file.

Introduce a queueCount_ to keep track of the requests queued.

When stopping the execution of the capture script the queueCount_ is not
reseted and the capture is continues as it is (i.e it is not stopped or
restarted).

Requests are queued with any controls the script matching the current
queueCount_.