qcam: Use QDialog for selection of cameras
Currently we use QInputDialog convenience dialogs to allow the user to
select a camera. This doesn't allow adding of more information (such as
camera location, model etc).

Create a QDialog with a QFormLayout that shows a QComboBox with camera
Ids. Use a QButtonBox to provide buttons for accepting and cancelling the
action.

Signed-off-by: Utkarsh Tiwari <utkarsh02t@gmail.com>