qcam: Support Hotplug for Camera Selection Dialog
Currently if there is HotPlug event when the user is on the Camera
selection dialog, the QComboBox didn't update to reflect the change.

If the cameraIdComboBox_ contains a valid QComboBox then update it to
reflect the HotPlug change. Check the validity of QComboBox pointer by
guarding it with a QPointer.

Signed-off-by: Utkarsh Tiwari <utkarsh02t@gmail.com>