qcam: CamSelectDialog: Add capture script button
Display a QLabel which is readonly, and displays the currently
selected capture script. The tooltip of the QLabel displays the file
path of the script.

Implement a capture script button which is a QToolButton which when
clicked opens a QFileDialog this allows to select a capture script
(*.yaml) file.

Next to the capture scipt button, show a QToolButton which stops the
capture script.

If an invalid script has been selected show a QMesssageBox::critical and
continue with the capture's previous state.

Introduce a queueCount_ to keep track of the requests queued.

When stopping the execution of the capture script the queueCount_ is not
reset and the capture continues as it is (i.e it is not stopped or
restarted).

Requests are queued with any controls from the script matching the current
queueCount_.

Signed-off-by: Utkarsh Tiwari <utkarsh02t@gmail.com>